Add Sunburst.AppxPackaging.PriAssemblyResolver project
This project contains code adapte from one of Microsoft's
Desktop Bridge samples. It can be used to load localized
satellite assemblies from an AppX package in a way that
allows unused languages to not be installed, and have the
app still function.

Note that this project doesn't yet compile, because it is
missing references. Those will be added over the next few commits.